Use `darkTheme` on import instead of import properties directly from tokens
It is a temporary change, justo to make application run again
Next step is remova all `tokens` import and controll all sttyles with CSS variables
