+++ src/components/CategoryBadge/index.js
@@ -11,7 +11,7 @@ const CategoryBadge = ({ className, size, color }) => (
   >
     <g>
       <path
-        fill={color}
+        fill={`var(--${color})`}
         d="M 69.14,27 56.25,4.71 A 9.43,9.43 0 0 0 48.1,0 H 22.3 A 9.43,9.43 0 0 0 14.15,4.71 L 1.26,27 a 9.43,9.43 0 0 0 0,9.42 l 12.89,22.37 a 9.45,9.45 0 0 0 8.15,4.71 h 25.8 a 9.45,9.45 0 0 0 8.15,-4.71 L 69.14,36.46 a 9.43,9.43 0 0 0 0,-9.46 z"
       />
     </g>
