+++ src/components/CategoryContainer/index.js
@@ -6,15 +6,15 @@ import categoriesInfo from "../../utils/categories-info"
 import * as S from "./styled"
 
 const CategoryContainer = ({ category }) => {
-  const { color, label, slug } = categoriesInfo(category)
+  const { cssVariable, label, slug } = categoriesInfo(category)
 
   return (
     <S.CategoryContainer
       to={`/${slug}`}
       title={`ver todos os post sobre ${label}`}
     >
-      <S.CategoryBadge size={16} color={color} />
-      <S.Category color={color}>{label}</S.Category>
+      <S.CategoryBadge size={16} color={cssVariable} />
+      <S.Category color={cssVariable}>{label}</S.Category>
     </S.CategoryContainer>
   )
 }
