+++ src/components/CategoryHeader/index.js
@@ -2,12 +2,12 @@ import React from "react"
 
 import * as S from "./styled"
 
-const CategoryHeader = ({ category: { color, description, label } }) => (
+const CategoryHeader = ({ category: { cssVariable, description, label } }) => (
   <S.Content>
     <S.Title>
       <S.CategoryContainer>
-        <S.CategoryBadge color={color} size={24} />
-        <S.Category color={color}>{label}</S.Category>
+        <S.CategoryBadge color={cssVariable} size={24} />
+        <S.Category color={cssVariable}>{label}</S.Category>
       </S.CategoryContainer>
     </S.Title>
