+++ src/components/Pagination/styled.js
@@ -1,8 +1,10 @@
 import styled from "styled-components"
 
-import { colors, fonts, radius, shadows } from "../../styles/tokens"
+import { darkTheme } from "../../styles/tokens"
 
 import Button from "../Button"
+
+const { colors, fonts, radius, shadows } = darkTheme;
 
 export const Container = styled.div`
   display: flex;
